api changes to functions for vectors and points

+++ session_cpp/src/vector.cpp
@@ -24,7 +24,7 @@ bool Vector::operator!=(const Vector &other) const { return !(*this == other); }
 // No-copy operators
 ///////////////////////////////////////////////////////////////////////////////////////////
 
-double &Vector::operator[](int index) {
+float &Vector::operator[](int index) {
   invalidate_length_cache();
   if (index == 0)
     return _x;
@@ -33,7 +33,7 @@ double &Vector::operator[](int index) {
   return _z; // assume index == 2
 }
 
-const double &Vector::operator[](int index) const {
+const float &Vector::operator[](int index) const {
   if (index == 0)
     return _x;
   if (index == 1)
@@ -41,14 +41,14 @@ const double &Vector::operator[](int index) const {
   return _z; // assume index == 2
 }
 
-Vector &Vector::operator*=(double factor) {
+Vector &Vector::operator*=(float factor) {
   set_x(_x * factor);
   set_y(_y * factor);
   set_z(_z * factor);
   return *this;
 }
 
-Vector &Vector::operator/=(double factor) {
+Vector &Vector::operator/=(float factor) {
   set_x(_x / factor);
   set_y(_y / factor);
   set_z(_z / factor);
@@ -73,9 +73,9 @@ Vector &Vector::operator-=(const Vector &other) {
 // Copy operators
 ///////////////////////////////////////////////////////////////////////////////////////////
 
-Vector Vector::operator*(double factor) const { return Vector(_x * factor, _y * factor, _z * factor); }
+Vector Vector::operator*(float factor) const { return Vector(_x * factor, _y * factor, _z * factor); }
 
-Vector Vector::operator/(double factor) const { return Vector(_x / factor, _y / factor, _z / factor); }
+Vector Vector::operator/(float factor) const { return Vector(_x / factor, _y / factor, _z / factor); }
 
 Vector Vector::operator+(const Vector &other) const {
   return Vector(_x + other._x, _y + other._y, _z + other._z);
@@ -85,7 +85,7 @@ Vector Vector::operator-(const Vector &other) const {
   return Vector(_x - other._x, _y - other._y, _z - other._z);
 }
 
-Vector operator*(double factor, const Vector &v) { return v * factor; }
+Vector operator*(float factor, const Vector &v) { return v * factor; }
 
 ///////////////////////////////////////////////////////////////////////////////////////////
 // JSON
@@ -124,9 +124,9 @@ Vector Vector::from_json(const std::string &filepath) {
 // Static methods
 ///////////////////////////////////////////////////////////////////////////////////////////
 
-Vector Vector::x_axis() { return Vector(1.0, 0.0, 0.0); }
-Vector Vector::y_axis() { return Vector(0.0, 1.0, 0.0); }
-Vector Vector::z_axis() { return Vector(0.0, 0.0, 1.0); }
+Vector Vector::x_axis() { return Vector(1.0f, 0.0f, 0.0f); }
+Vector Vector::y_axis() { return Vector(0.0f, 1.0f, 0.0f); }
+Vector Vector::z_axis() { return Vector(0.0f, 0.0f, 1.0f); }
 
 Vector Vector::from_start_and_end(const Vector &start, const Vector &end) {
   return Vector(end._x - start._x, end._y - start._y, end._z - start._z);
@@ -143,19 +143,19 @@ void Vector::reverse() {
   // Length magnitude stays the same, no need to invalidate cache
 }
 
-double Vector::compute_length() const {
-  double len = 0.0;
+float Vector::compute_length() const {
+  float len = 0.0f;
 
-  double ax = std::abs(_x);
-  double ay = std::abs(_y);
-  double az = std::abs(_z);
+  float ax = std::abs(_x);
+  float ay = std::abs(_y);
+  float az = std::abs(_z);
 
   const bool x_zero = ax < geo::GLOBALS::ZERO_TOLERANCE;
   const bool y_zero = ay < geo::GLOBALS::ZERO_TOLERANCE;
   const bool z_zero = az < geo::GLOBALS::ZERO_TOLERANCE;
 
   if (x_zero && y_zero && z_zero)
-    return 0.0;
+    return 0.0f;
   else if (x_zero && y_zero)
     return az;
   else if (x_zero && z_zero)
@@ -173,17 +173,17 @@ double Vector::compute_length() const {
   if (ax > geo::GLOBALS::DOUBLE_MIN) {
     ay /= ax;
     az /= ax;
-    len = ax * std::sqrt(1.0 + ay * ay + az * az);
-  } else if (ax > 0.0 && geo::GLOBALS::IS_FINITE(ax)) {
+    len = ax * std::sqrt(1.0f + ay * ay + az * az);
+  } else if (ax > 0.0f && geo::GLOBALS::IS_FINITE(ax)) {
     len = ax;
   } else {
-    len = 0.0;
+    len = 0.0f;
   }
 
   return len;
 }
 
-double Vector::cached_length() const {
+float Vector::cached_length() const {
   if (!_has_length) {
     _length = compute_length();
     _has_length = true;
@@ -191,15 +191,15 @@ double Vector::cached_length() const {
   return _length;
 }
 
-double Vector::length() const { return cached_length(); }
+float Vector::length() const { return cached_length(); }
 
-double Vector::length_squared() const {
+float Vector::length_squared() const {
   return _x * _x + _y * _y + _z * _z;
 }
 
 bool Vector::unitize() {
-  double d = compute_length();
-  if (d > 0.0) {
+  float d = compute_length();
+  if (d > 0.0f) {
     set_x(_x / d);
     set_y(_y / d);
     set_z(_z / d);
@@ -214,12 +214,12 @@ Vector Vector::unitized() {
   return u;
 }
 
-std::tuple<Vector, double, Vector, double>
-Vector::projection(Vector &projection_vector, double tolerance) {
-  double projection_vector_length = projection_vector.length();
+std::tuple<Vector, float, Vector, float>
+Vector::projection(Vector &projection_vector, float tolerance) {
+  float projection_vector_length = projection_vector.length();
 
   if (projection_vector_length < tolerance) {
-    return {Vector(0, 0, 0), 0.0, Vector(0, 0, 0), 0.0};
+    return {Vector(0, 0, 0), 0.0f, Vector(0, 0, 0), 0.0f};
   }
 
   Vector projection_vector_unit(
@@ -227,11 +227,11 @@ Vector::projection(Vector &projection_vector, double tolerance) {
       projection_vector._y / projection_vector_length,
       projection_vector._z / projection_vector_length);
 
-  double projected_vector_length = this->dot(projection_vector_unit);
+  float projected_vector_length = this->dot(projection_vector_unit);
   Vector out_projection_vector = projection_vector_unit * projected_vector_length;
 
   Vector out_perpendicular_projected_vector = *this - out_projection_vector;
-  double out_perpendicular_projected_vector_length = out_perpendicular_projected_vector.length();
+  float out_perpendicular_projected_vector_length = out_perpendicular_projected_vector.length();
 
   return {out_projection_vector,
           projected_vector_length,
@@ -240,14 +240,14 @@ Vector::projection(Vector &projection_vector, double tolerance) {
 }
 
 int Vector::is_parallel_to(const Vector &other) {
-  double ll = cached_length() * other.cached_length();
+  float ll = cached_length() * other.cached_length();
   int result;
   
-  if (ll > 0.0) {
-    const double cos_angle = ((*this)[0] * other[0] + (*this)[1] * other[1] + (*this)[2] * other[2]) / ll;
+  if (ll > 0.0f) {
+    const float cos_angle = ((*this)[0] * other[0] + (*this)[1] * other[1] + (*this)[2] * other[2]) / ll;
     
-    const double angle_in_radians = geo::GLOBALS::ANGLE * (geo::GLOBALS::PI / 180.0);
-    const double cos_tol = std::cos(angle_in_radians);
+    const float angle_in_radians = geo::GLOBALS::ANGLE * (geo::GLOBALS::PI / 180.0f);
+    const float cos_tol = std::cos(angle_in_radians);
     if (cos_angle >= cos_tol)
       result = 1;  // Parallel
     else if (cos_angle <= -cos_tol)
@@ -261,8 +261,8 @@ int Vector::is_parallel_to(const Vector &other) {
   return result;
 }
 
-double Vector::dot(const Vector &other) {
-  double result = 0.0;
+float Vector::dot(const Vector &other) {
+  float result = 0.0f;
   for (size_t i = 0; i < 3; ++i) {
     result += (*this)[i] * other[i];
   }
@@ -270,60 +270,60 @@ double Vector::dot(const Vector &other) {
 }
 
 Vector Vector::cross(const Vector &other) {
-  double cx = (*this)[1] * other[2] - (*this)[2] * other[1];
-  double cy = (*this)[2] * other[0] - (*this)[0] * other[2];
-  double cz = (*this)[0] * other[1] - (*this)[1] * other[0];
+  float cx = (*this)[1] * other[2] - (*this)[2] * other[1];
+  float cy = (*this)[2] * other[0] - (*this)[0] * other[2];
+  float cz = (*this)[0] * other[1] - (*this)[1] * other[0];
   Vector result(cx, cy, cz);
   result.unitize();
   return result;
 }
 
-double Vector::angle(const Vector &other, bool sign_by_cross_product, bool degrees,
-                     double tolerance) {
-  double dotp = this->dot(other);
-  double len0 = this->cached_length();
-  double len1 = other.cached_length();
-  double denom = len0 * len1;
+float Vector::angle(const Vector &other, bool sign_by_cross_product, bool degrees,
+                     float tolerance) {
+  float dotp = this->dot(other);
+  float len0 = this->cached_length();
+  float len1 = other.cached_length();
+  float denom = len0 * len1;
   if (denom < tolerance) {
-    return 0.0;
+    return 0.0f;
   }
-  double cos_angle = dotp / denom;
-  cos_angle = std::max(-1.0, std::min(1.0, cos_angle));
-  double ang = std::acos(cos_angle);
+  float cos_angle = dotp / denom;
+  cos_angle = std::max(-1.0f, std::min(1.0f, cos_angle));
+  float ang = std::acos(cos_angle);
   if (sign_by_cross_product) {
     Vector cp = this->cross(other);
     if (cp._z < 0)
       ang = -ang;
   }
-  double to_degrees = degrees ? geo::GLOBALS::TO_DEGREES : 1.0;
+  float to_degrees = degrees ? geo::GLOBALS::TO_DEGREES : 1.0f;
   return ang * to_degrees;
 }
 
-Vector Vector::get_leveled_vector(double &vertical_height) {
+Vector Vector::get_leveled_vector(float &vertical_height) {
   Vector copy(_x, _y, _z);
   if (copy.unitize()) {
     Vector reference(0, 0, 1);
-    double angle = copy.angle(reference, true); // returns degrees
+    float angle = copy.angle(reference, true); // returns degrees
     // CRITICAL: statics bug - passes degrees directly to cos (expects radians)
-    double inclined_offset_by_vertical_distance = vertical_height / std::cos(angle);
+    float inclined_offset_by_vertical_distance = vertical_height / std::cos(angle);
     copy *= inclined_offset_by_vertical_distance;
   }
   return copy;
 }
 
-double Vector::cosine_law(double &a, double &b, double &ang_between, bool degrees) {
-  double to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0;
+float Vector::cosine_law(float &a, float &b, float &ang_between, bool degrees) {
+  float to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0f;
   return std::sqrt(a * a + b * b - 2 * a * b * std::cos(ang_between * to_rad));
 }
 
-double Vector::sine_law_angle(double &a, double &A, double &b, bool degrees) {
-  double to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0;
-  double to_deg = degrees ? geo::GLOBALS::TO_DEGREES : 1.0;
+float Vector::sine_law_angle(float &a, float &A, float &b, bool degrees) {
+  float to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0f;
+  float to_deg = degrees ? geo::GLOBALS::TO_DEGREES : 1.0f;
   return std::asin((b * std::sin(A * to_rad)) / a) * to_deg;
 }
 
-double Vector::sine_law_length(double &a, double &A, double &B, bool degrees) {
-  double to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0;
+float Vector::sine_law_length(float &a, float &A, float &B, bool degrees) {
+  float to_rad = degrees ? geo::GLOBALS::TO_RADIANS : 1.0f;
   return (a * std::sin(B * to_rad)) / std::sin(A * to_rad);
 }
 
@@ -332,7 +332,7 @@ double Vector::angle_between_vector_xy_components(Vector &vector) {
 }
 
 Vector Vector::sum_of_vectors(std::vector<Vector> &vectors) {
-  double sx = 0, sy = 0, sz = 0;
+  float sx = 0, sy = 0, sz = 0;
   for (const auto &v : vectors) {
     sx += v[0];
     sy += v[1];
@@ -341,52 +341,52 @@ Vector Vector::sum_of_vectors(std::vector<Vector> &vectors) {
   return Vector(sx, sy, sz);
 }
 
-std::array<double, 3> Vector::coordinate_direction_3angles(bool degrees) {
-  double x_coord = _x;
-  double y_coord = _y;
-  double z_coord = _z;
-  double r = std::sqrt(x_coord * x_coord + y_coord * y_coord + z_coord * z_coord);
+std::array<float, 3> Vector::coordinate_direction_3angles(bool degrees) {
+  float x_coord = _x;
+  float y_coord = _y;
+  float z_coord = _z;
+  float r = std::sqrt(x_coord * x_coord + y_coord * y_coord + z_coord * z_coord);
   
   if (r == 0) {
     return {0, 0, 0};
   }
   
   // unit vector proportions
-  double x_proportion = x_coord / r;
-  double y_proportion = y_coord / r;
-  double z_proportion = z_coord / r;
+  float x_proportion = x_coord / r;
+  float y_proportion = y_coord / r;
+  float z_proportion = z_coord / r;
   
   // angles
-  double alpha = std::acos(x_proportion);
-  double beta = std::acos(y_proportion);
-  double gamma = std::acos(z_proportion);
+  float alpha = std::acos(x_proportion);
+  float beta = std::acos(y_proportion);
+  float gamma = std::acos(z_proportion);
   
   if (degrees) {
-    alpha = alpha * 180.0 / geo::GLOBALS::PI;
-    beta = beta * 180.0 / geo::GLOBALS::PI;
-    gamma = gamma * 180.0 / geo::GLOBALS::PI;
+    alpha = alpha * 180.0f / geo::GLOBALS::PI;
+    beta = beta * 180.0f / geo::GLOBALS::PI;
+    gamma = gamma * 180.0f / geo::GLOBALS::PI;
   }
   
   return {alpha, beta, gamma};
 }
 
-std::array<double, 2> Vector::coordinate_direction_2angles(bool degrees) {
-  double x_coord = _x;
-  double y_coord = _y;
-  double z_coord = _z;
-  double r = std::sqrt(x_coord * x_coord + y_coord * y_coord + z_coord * z_coord);
+std::array<float, 2> Vector::coordinate_direction_2angles(bool degrees) {
+  float x_coord = _x;
+  float y_coord = _y;
+  float z_coord = _z;
+  float r = std::sqrt(x_coord * x_coord + y_coord * y_coord + z_coord * z_coord);
   
   if (r == 0) {
     return {0, 0};
   }
   
   // spherical coordinates
-  double phi = std::acos(z_coord / r);
-  double theta = std::atan2(y_coord, x_coord);
+  float phi = std::acos(z_coord / r);
+  float theta = std::atan2(y_coord, x_coord);
   
   if (degrees) {
-    phi = phi * 180.0 / geo::GLOBALS::PI;
-    theta = theta * 180.0 / geo::GLOBALS::PI;
+    phi = phi * 180.0f / geo::GLOBALS::PI;
+    theta = theta * 180.0f / geo::GLOBALS::PI;
   }
   
   return {phi, theta};
@@ -394,7 +394,7 @@ std::array<double, 2> Vector::coordinate_direction_2angles(bool degrees) {
 
 bool Vector::perpendicular_to(Vector &v) {
   int i, j, k;
-  double a, b;
+  float a, b;
   k = 2;
   if (std::fabs(v[1]) > std::fabs(v[0])) {
     if (std::fabs(v[2]) > std::fabs(v[1])) {
@@ -418,17 +418,17 @@ bool Vector::perpendicular_to(Vector &v) {
     i = 0; j = 1; k = 2; a = v[0]; b = -v[1];
   }
 
-  double arr[3] = {_x, _y, _z};
+  float arr[3] = {_x, _y, _z};
   arr[i] = b;
   arr[j] = a;
-  arr[k] = 0.0;
+  arr[k] = 0.0f;
   set_x(arr[0]);
   set_y(arr[1]);
   set_z(arr[2]);
-  return (a != 0.0) ? true : false;
+  return (a != 0.0f) ? true : false;
 }
 
-void Vector::scale(double factor) {
+void Vector::scale(float factor) {
   set_x(_x * factor);
   set_y(_y * factor);
   set_z(_z * factor);
@@ -436,7 +436,7 @@ void Vector::scale(double factor) {
 
 void Vector::scale_up() { scale(geo::GLOBALS::SCALE); }
 
-void Vector::scale_down() { scale(1.0 / geo::GLOBALS::SCALE); }
+void Vector::scale_down() { scale(1.0f / geo::GLOBALS::SCALE); }
 
 ///////////////////////////////////////////////////////////////////////////////////////////
 // Not class methods
